rename similar to quests-sets
